реализуем абстрактный класс
