+++ src/view/Abstract-view.js
@@ -1,0 +1,27 @@
+import { createElement } from '../render';
+
+export default class AbstractView {
+  #element = null;
+
+  constructor() {
+    if (new.target === AbstractView) {
+      throw new Error('Can\'t instantiate AbstractView');
+    }
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  get template() {
+    throw new Error('Abstract method not impleted: get template');
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
