+++ src/view/site-list-view.js
@@ -1,26 +1,12 @@
-import { createElement } from '../render';
+import AbstractView from './Abstract-view';
 
 const createEventsListTemplate = () => (
   `<ul class="trip-events__list">
    </ul>`
 );
 
-export default class EventsListTemplate {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class EventsListTemplate extends AbstractView {
   get template() {
     return createEventsListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
